feat: implement permissions to controller

+++ Modules/Kandang/app/Http/Controllers/MasterData/KandangController.php
@@ -4,6 +4,7 @@ namespace Modules\Kandang\Http\Controllers\MasterData;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Modules\Kandang\Models\Kandang;
 
 class KandangController extends Controller
@@ -14,6 +15,8 @@ class KandangController extends Controller
 
     public function index()
     {
+        Gate::authorize('Lihat Semua Kandang');
+
         $datas = $this->kandang
             ->query()
             ->when(request()->query('search'), function($query, $search) {
@@ -21,6 +24,7 @@ class KandangController extends Controller
             })
             ->orderBy('created_at', 'desc')
             ->paginate(request()->get('perPage', 10));
+
         return view('kandang::master-data.kandang.index', compact('datas'));
     }
 
@@ -29,6 +33,8 @@ class KandangController extends Controller
      */
     public function create()
     {
+        Gate::authorize('Tambah Kandang');
+
         return view('kandang::master-data.kandang.create');
     }
 
@@ -36,6 +42,8 @@ class KandangController extends Controller
      * Store a newly created resource in storage.
      */
     public function store(Request $request) {
+        Gate::authorize('Tambah Kandang');
+
         $request->validate([
             'nama' => ['required', 'string', 'max:255'],
             'alamat' => ['required', 'string', 'max:1000'],
@@ -51,7 +59,10 @@ class KandangController extends Controller
      */
     public function edit($id)
     {
+        Gate::authorize('Edit Kandang');
+
         $data = $this->kandang->findOrFail($id);
+
         return view('kandang::master-data.kandang.edit', compact('data'));
     }
 
@@ -59,6 +70,8 @@ class KandangController extends Controller
      * Update the specified resource in storage.
      */
     public function update(Request $request, $id) {
+        Gate::authorize('Edit Kandang');
+
         $request->validate([
             'nama' => ['required', 'string', 'max:255'],
             'alamat' => ['required', 'string', 'max:1000'],
@@ -76,6 +89,8 @@ class KandangController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy($id) {
+        Gate::authorize('Hapus Kandang');
+
         $kandang = $this->kandang->findOrFail($id);
         $kandang->delete();
